Added adjusted p-values to VEGAS results
The reading routine now adds a column with adjusted gene-wise p-values to
the raw VEGAS results (default: Bonferroni-Holm). These adjusted p-values
are carried forward through all manipulations and exports, and are the
criterion for the topTable cutoff. Note that when the gene list is
shortened (in a non-informative manner, ie when dropping duplicates or
intersecting between data sets), the adjustment is updated.