Changed CIs for shared genes to binomial
For the function counts(), the confidence intervals were flipped from
normal apporximations for the number expected under the null to binomial
confidence intervals for the observed count. Here, the expected count
under the null is assumed known and determined by the product of the
marginal proportion of signficant genes in each list, i.e. we condition on
the marginal distributions. Given the application, it doesn't need to be
more kosher than this (hopefully).